The 4.15.0-rc5 ! Tinkerboard friendly !
I took the time to test it on a real Tinkerboard, this time and
deal with bug #1 . So the endless stream of mipi-dsi and rockchip-drm
errors is no more !

...

And so is the mipi-dsi node, which is clearly not the way to solve
the issue.
I guess I'll have to inspect this a bit more when I'll acquire a MIPI
touch screen.

Anyway, it works on HDMI, GPU drivers were tested correctly.

Signed-off-by: Myy Miouyouyou <myy@miouyouyou.fr>